+++ Assets/Project/Scripts/Environment/Global/IGlobalCoverageTimerObserver.cs
@@ -4,6 +4,6 @@ namespace Invaders.Environment.Global
 {
     public interface IGlobalCoverageTimerObserver
     {
-        event Action<int, int> OnStepInterval;
+        event Action<float, float> OnStepInterval;
     }
 }
